+++ tests/base.test.js
@@ -1910,9 +1910,12 @@ cellType onoff
  enum on off
 nodeType error
  baseNodeType errorNode
-abstract topLevel
-abstract someAbstractClass
-abstract colorProperties
+nodeType topLevel
+ abstract
+nodeType someAbstractClass
+ abstract
+nodeType colorProperties
+ abstract
  extends topLevel
  cells int
 nodeType hue